add streaming support
this adds streaming support for grenache.

the server emits a `stream` event before the request is
parsed and buffered. with the setting `disableBuffered: true`
we can then disable the convenience auto-parsing and allow more
granular control on the serverside with the req and response
objects / streams. an object called `meta` is emitted, with
`isStream (bool)`, indicating a streaming client. in case
of a stream it also contains req-id and key, which are parsed
from the http header, as the stream is reserved for the (binary)
data.

when a stream is sent with the `strean` method, it
will add the grenache-specific req-id and key to the request
headers. if we want to send additional, grenache-specific data,
the arguments and actions, we can add those in our request:

```js
const req = pst.stream('rpc_stream', {
  timeout: 10000,
  headers: {
    _a: 'uploadPublic',
    _ar: { foo: 'bar' }
  }
})
```
closes: https://github.com/bitfinexcom/grenache-nodejs-http/issues/21